SVG: Implement handling of <animate>, <animateTransform> and <animateMotion> SVG tags (see notes below)
This is not a full animation implementation, we simply delegate the animated attributes to the parent tags,
allowing us to use these attributes in our static rendering of the SVG nodes.
We currently handle the following attributes in the following notes:

- 'd' attribute is added to <Path> parent
- 'cx', 'cy', and 'r' attributes are added to <Circle> parent
- 'transform' attribute is added to <Use> parent

These attributes are only set in the parent, if the parent does not already have values set for them.

TODO: Add more cases as we happen upon files that have issues.